Build: Remove tmp/ directory only if present and make Travis CI :happy_cat:

+++ test/integration/compilers_test.rb
@@ -4,7 +4,7 @@ describe 'Compilers' do
   before do
     root = __dir__ + '/../fixtures'
     tmp  = Pathname.new(__dir__ + '/../../tmp')
-    tmp.rmtree
+    tmp.rmtree if tmp.exist?
 
     Lotus::Assets.configure do
       compile     true
